refactor: improve WebSocket chunking and transport limits
- Introduced a new `constants.ts` file to define shared transport limits for the Cloudflare sync provider, including `MAX_WS_MESSAGE_BYTES` and `MAX_PULL_EVENTS_PER_MESSAGE`.
- Refactored the chunking logic by moving `splitChunkBySize` to a new `ws-chunking.ts` file, enhancing clarity and reusability.
- Removed obsolete chunking logic from `shared.ts` and updated relevant files to utilize the new chunking implementation.
- Streamlined the push and pull operations in the WebSocket transport to ensure compliance with the new chunking limits.

These changes enhance the maintainability and clarity of the WebSocket transport logic, ensuring better handling of message sizes and chunking behavior.
